requirements.txt
absl-py==0.1.10
aiohttp==3.5.1
aiomysql==0.0.20
amqp==2.3.2
aniso8601==1.2.0
APScheduler==3.3.1
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.6.3
billiard==3.5.0.5
boto==2.49.0
boto3==1.9.74
botocore==1.12.74
bs4==0.0.1
bz2file==0.98
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
chromedriver==2.24.1
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==2.4.1
cssselect==1.0.3
DBUtils==1.3
decorator==4.3.0
Django==2.2.7
django-ckeditor==5.8.0
django-cors-headers==2.4.0
django-js-asset==1.2.2
docopt==0.6.2
docutils==0.14
Flask==0.12.1
Flask-BasicAuth==0.2.0
Flask-RESTful==0.3.5
flask-restful-swagger==0.19
Flask-SQLAlchemy==2.2
gensim==3.6.0
gerapy==0.8.5
hyperlink==18.0.0
idna==2.7
idna-ssl==1.1.0
incremental==17.5.0
ipython==7.1.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.13.1
jieba==0.39
Jinja2==2.9.6
jmespath==0.9.3
kombu==4.2.2.post1
lxml==4.2.5
MarkupSafe==1.0
more-itertools==4.3.0
msgpack-python==0.5.6
multidict==4.5.2
mysqlclient==1.4.2.post1
numpy==1.15.2
panda==0.3.1
pandas==0.23.4
parsel==1.5.1
parso==0.3.1
pdfminer3k==1.3.1
pep8==1.7.1
pickleshare==0.7.5
pika==0.12.0
Pillow==5.3.0
Pillow-PIL==0.1.dev0
pluggy==0.7.1
ply==3.11
prompt-toolkit==2.0.7
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
pymongo==3.7.1
pymssql==2.1.4
PyMySQL==0.9.3
pyOpenSSL==18.0.0
PySocks==1.6.8
pytest==3.8.2
python-dateutil==2.6.0
python-docx==0.8.7
python-scrapyd-api==2.1.2
pytz==2018.9
PyYAML==5.1
render==1.0.0
requests==2.21.0
Scrapy==1.5.1
six==1.12.0
SpiderKeeper==1.2.0
SQLAlchemy==1.1.9
sqlparse==0.3.0
urllib3==1.24.1
virtualenv==16.7.7
Werkzeug==0.12.1
